requirements.txt
streamlit>=1.36.0
ultralytics>=8.1.0
opencv-python
torch>=2.0.0
torchvision
numpy
pandas
Pillow
